feat(workflow): add auto layout to workflow
