+++ packages/app-builder/src/components/Scenario/Workflow/DetailPannel/AddToCaseIfPossibleNode.tsx
@@ -1,46 +1,25 @@
 import { Callout } from '@app-builder/components/Callout';
 import { Highlight } from '@app-builder/components/Highlight';
 import { type Inbox } from '@app-builder/models/inbox';
-import { Await } from '@remix-run/react';
 import { matchSorter } from 'match-sorter';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Input, SelectWithCombobox } from 'ui-design-system';
 
 import { type AddToCaseIfPossibleAction } from '../models/node-data';
-import { useWorkflowActions } from '../store';
 import { workflowI18n } from '../workflow-i18n';
-import { useWorkflowData } from '../WorkflowData';
+import { useWorkflowActions, useWorkflowData } from '../WorkflowProvider';
 import { defaultCaseName } from './shared';
 
-export function AddToCaseIfPossibleNode(props: {
+export function AddToCaseIfPossibleNode({
+  id,
+  data,
+}: {
   id: string;
   data: AddToCaseIfPossibleAction;
 }) {
   const { t } = useTranslation(workflowI18n);
   const { inboxes } = useWorkflowData();
-
-  return (
-    <React.Suspense fallback={t('common:loading')}>
-      <Await resolve={inboxes}>
-        {(inboxes) => (
-          <AddToCaseIfPossibleNodeImpl inboxes={inboxes} {...props} />
-        )}
-      </Await>
-    </React.Suspense>
-  );
-}
-
-function AddToCaseIfPossibleNodeImpl({
-  id,
-  data,
-  inboxes,
-}: {
-  id: string;
-  data: AddToCaseIfPossibleAction;
-  inboxes: Inbox[];
-}) {
-  const { t } = useTranslation(workflowI18n);
   const { updateNode } = useWorkflowActions();
 
   return (
