+++ packages/app-builder/src/components/Scenario/Workflow/Nodes/AddToCaseIfPossibleActionContent.tsx
@@ -1,11 +1,9 @@
-import { type Inbox } from '@app-builder/models/inbox';
-import { Await } from '@remix-run/react';
 import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 
 import { type AddToCaseIfPossibleAction } from '../models/node-data';
 import { workflowI18n } from '../workflow-i18n';
-import { useWorkflowData } from '../WorkflowData';
+import { useWorkflowData } from '../WorkflowProvider';
 
 export function AddToCaseIfPossibleActionContent({
   data,
@@ -14,31 +12,6 @@ export function AddToCaseIfPossibleActionContent({
 }) {
   const { t } = useTranslation(workflowI18n);
   const { inboxes } = useWorkflowData();
-
-  return (
-    <div className="flex flex-col gap-1">
-      <React.Suspense fallback={t('common:loading')}>
-        <Await resolve={inboxes}>
-          {(inboxes) => (
-            <AddToCaseIfPossibleActionContentImpl
-              data={data}
-              inboxes={inboxes}
-            />
-          )}
-        </Await>
-      </React.Suspense>
-    </div>
-  );
-}
-
-function AddToCaseIfPossibleActionContentImpl({
-  data,
-  inboxes,
-}: {
-  inboxes: Inbox[];
-  data: AddToCaseIfPossibleAction;
-}) {
-  const { t } = useTranslation(workflowI18n);
   const selectedInbox = React.useMemo(() => {
     if (!data.inboxId || !inboxes) return undefined;
     return inboxes.find((inbox) => inbox.id === data.inboxId);
@@ -46,22 +19,26 @@ function AddToCaseIfPossibleActionContentImpl({
 
   if (!selectedInbox) {
     return (
-      <p className="max-w-64 whitespace-pre-wrap">
-        {t('workflows:action_node.add_to_case_if_possible.empty_content')}
-      </p>
+      <div className="flex flex-col gap-1">
+        <p className="max-w-64 whitespace-pre-wrap">
+          {t('workflows:action_node.add_to_case_if_possible.empty_content')}
+        </p>
+      </div>
     );
   }
 
   return (
-    <p className="max-w-64 whitespace-pre-wrap">
-      <Trans
-        t={t}
-        i18nKey="workflows:action_node.add_to_case_if_possible.content"
-        components={{
-          Inbox: <span className="font-bold" />,
-        }}
-        values={{ inbox: selectedInbox.name }}
-      />
-    </p>
+    <div className="flex flex-col gap-1">
+      <p className="max-w-64 whitespace-pre-wrap">
+        <Trans
+          t={t}
+          i18nKey="workflows:action_node.add_to_case_if_possible.content"
+          components={{
+            Inbox: <span className="font-bold" />,
+          }}
+          values={{ inbox: selectedInbox.name }}
+        />
+      </p>
+    </div>
   );
 }
