+++ packages/app-builder/src/components/Scenario/Workflow/Nodes/DecisionCreatedTriggerContent.tsx
@@ -1,12 +1,10 @@
 import { Outcome } from '@app-builder/components/Decisions';
-import { type Scenario } from '@app-builder/models/scenario';
-import { Await } from '@remix-run/react';
 import * as React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 
 import { type DecisionCreatedTrigger } from '../models/node-data';
 import { workflowI18n } from '../workflow-i18n';
-import { useWorkflowData } from '../WorkflowData';
+import { useWorkflowData } from '../WorkflowProvider';
 
 export function DecisionCreatedTriggerContent({
   data,
@@ -15,31 +13,6 @@ export function DecisionCreatedTriggerContent({
 }) {
   const { t } = useTranslation(workflowI18n);
   const { scenarios } = useWorkflowData();
-
-  return (
-    <div className="flex flex-col gap-1">
-      <React.Suspense fallback={t('common:loading')}>
-        <Await resolve={scenarios}>
-          {(scenarios) => (
-            <DecisionCreatedTriggerContentImpl
-              data={data}
-              scenarios={scenarios}
-            />
-          )}
-        </Await>
-      </React.Suspense>
-    </div>
-  );
-}
-
-function DecisionCreatedTriggerContentImpl({
-  data,
-  scenarios,
-}: {
-  scenarios: Scenario[];
-  data: DecisionCreatedTrigger;
-}) {
-  const { t } = useTranslation(workflowI18n);
   const selectedScenario = React.useMemo(() => {
     if (!data.scenarioId || !scenarios) return undefined;
     return scenarios.find((scenario) => scenario.id === data.scenarioId);
@@ -47,14 +20,16 @@ function DecisionCreatedTriggerContentImpl({
 
   if (!selectedScenario || data.outcomes.length === 0) {
     return (
-      <p className="max-w-64 whitespace-pre-wrap">
-        {t('workflows:trigger_node.decision_created.empty_content')}
-      </p>
+      <div className="flex flex-col gap-1">
+        <p className="max-w-64 whitespace-pre-wrap">
+          {t('workflows:trigger_node.decision_created.empty_content')}
+        </p>
+      </div>
     );
   }
 
   return (
-    <>
+    <div className="flex flex-col gap-1">
       <p className="max-w-64 whitespace-pre-wrap">
         <Trans
           t={t}
@@ -70,6 +45,6 @@ function DecisionCreatedTriggerContentImpl({
           <Outcome key={outcome} outcome={outcome} border="square" size="big" />
         ))}
       </p>
-    </>
+    </div>
   );
 }
